some update to get to edit function on the users list for admin, not finished
