+++ templates/admin/users.html.twig
@@ -8,6 +8,7 @@
     <table class="table table-striped table-bordered border-dark">
         <thead class="table-secondary">
             <tr>
+                <th scope="col">Action</th>
                 <th scope="col">ID</th>
                 <th scope="col">Prénom</th>
                 <th scope="col">Nom</th>
@@ -20,6 +21,7 @@
         <tbody>
             {% for user in users %}
             <tr>
+                <th {#href="{{ path('admin_user_edit') }}"#}>Modifier</th>
                 <th scope="row">{{user.id}}</th>
                 <td>{{user.firstname}}</td>
                 <td>{{user.lastname}}</td>
